+++ voip_audio_route_manager_platform_interface/lib/voip_audio_route_manager_platform_interface.dart
@@ -1,3 +1,4 @@
 export 'src/audio_device_model.dart';
+export 'src/audio_route_result.dart';
 export 'src/platform_interface.dart';
 export 'src/method_channel_implementation.dart';